requirements.txt
albumentations==1.3.0
Cython==0.29.32
dicomsdl==0.109.1
efficientnet-pytorch==0.7.1
einops==0.6.1
gdown==4.7.1
glob2==0.7
huggingface-hub==0.10.0
imageio==2.22.1
imageio-ffmpeg==0.4.8
imgaug==0.4.0
ipykernel==6.16.0
iterative-stratification==0.1.7
joblib==1.2.0
kaggle==1.5.12
matplotlib==3.6.0
monai==1.0.1
nibabel==4.0.2
numpy==1.23.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.5.0
pretrainedmodels==0.7.4
pydicom==2.4.3
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.1
python-gdcm==3.0.22
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
segmentation-models-pytorch==0.3.3
transformers==4.23.1
timm==0.9.2
torch==1.12.0+cu113
torchaudio==0.12.0+rocm5.1.1
torchvision==0.13.0+cu113
tqdm==4.64.1
